feat(spring-boot): 05_repositorios_persistencia

+++ src/main/java/ec/edu/ups/icc/fundamentos01/products/dtos/UpdateProductsDto.java
@@ -1,7 +1,9 @@
 package ec.edu.ups.icc.fundamentos01.products.dtos;
 
+import java.math.BigDecimal;
+
 public class UpdateProductsDto {
     public String name;
-    public Double price;
+    public BigDecimal price;
     public Integer stock;
 }
